refactor(core): extract shared provider helpers from bootstrap (#2916)
Move BootstrapError, build_provider_for_switch, build_provider_from_entry,
and effective_embedding_model from crates/zeph-core/src/bootstrap/ into a
new internal module crates/zeph-core/src/provider_factory.rs.

Update all 5 internal callers (autodream, arise, magic_docs, provider_cmd,
session_config) to use crate::provider_factory::. Add pub use re-exports in
bootstrap/provider.rs and bootstrap/skills.rs so the binary crate is
unaffected. create_named_provider and create_summary_provider remain in
bootstrap (no internal agent callers).

Prerequisite for moving bootstrap/ to the binary crate (closes #2906 pre-work).

Closes #2916
